with  final touches

+++ src/main/java/com/example/agile/AgileApplication.java
@@ -36,7 +36,7 @@ public class AgileApplication {
 			Set<Role> roles = new HashSet<>();
 			roles.add(adminRole);
 
-			ApplicationUser admin = new ApplicationUser(1, "admin", passwordEncoder.encode("password"), roles);
+			ApplicationUser admin = new ApplicationUser((long) 1, "admin", passwordEncoder.encode("password"), roles);
 
 		  };
 		}
