+++ src/main/java/com/example/agile/services/AuthenticationService.java
@@ -45,7 +45,7 @@ public class AuthenticationService {
         Set<Role> authorities = new HashSet<>();
         authorities.add(userRole);
 
-        return userRepository.save(new ApplicationUser(0, username, encodedPassword, authorities));
+        return userRepository.save(new ApplicationUser((long) 0, username, encodedPassword, authorities));
     }
